IMPLEMENTED Put and Delete Mapping restriction
Put and Delete Mapping is only allowed for the resource of the
authorized user and not for the other resources

+++ src/main/java/com/rastatech/secretrasta/controller/VoteController.java
@@ -3,12 +3,15 @@ package com.rastatech.secretrasta.controller;
 import com.rastatech.secretrasta.dto.WishVoteRequest;
 import com.rastatech.secretrasta.dto.WishVoteResponse;
 import com.rastatech.secretrasta.model.WishVoteEntity;
+import com.rastatech.secretrasta.repository.WishVoteRepository;
 import com.rastatech.secretrasta.service.UserService;
 import com.rastatech.secretrasta.service.WishVoteService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -22,6 +25,7 @@ public class VoteController {
     private final ModelMapper modelMapper;
     private final WishVoteService wishVoteService;
     private final UserService userService;
+    private final WishVoteRepository wishVoteRepository;
 
     @PostMapping("/{wish_id}")
     public void vote(Authentication auth,
@@ -34,7 +38,10 @@ public class VoteController {
 
     @DeleteMapping("/{vote_id}")
     public void deleteVote(@PathVariable("wish_id") Long wishId,
-                           @PathVariable("vote_id") Long voteId) {
+                           @PathVariable("vote_id") Long voteId,
+                           Authentication auth) {
+        if (!auth.getPrincipal().equals(wishVoteRepository.findById(voteId).get().getUser().getUsername()))
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
         wishVoteService.deleteVote(wishId, voteId);
     }
